Renamed seed data. Consolidated into one folder in the server folder.
