+++ server/Seed Data/seed.js
@@ -1,13 +1,13 @@
 // seed.js
-//TODO: review location of seed file
 
-const { connectionString } = require('./index.js');
+
+const { connectionString } = require('../models/index.js');
 
 const mongoose = require('mongoose');
-const Business = require('./business');
-const Categories = require('./categories');
-const seedData = require('./mockBusiness'); //TODO: rename to seedBusiness / review location of mock data
-const seedCategories = require('./mockCategories.js'); //TODO: review location of mock data
+const Business = require('../models/business.js');
+const Categories = require('../models/categories.js');
+const seedData = require('./seedBusinesses.js'); 
+const seedCategories = require('./seedCategories.js'); 
 
 //FIXME: do we need to remove the console logs here -> if its only use for dev?
 const seedDB = async () => {
